Added basic non-functional/non-tested mon health check
