+++ pkg/operator/cluster/ceph/mon/nodes.go
@@ -1,38 +1,20 @@
-/*
-Copyright 2018 The Rook Authors. All rights reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-	http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-// Package mon for the Ceph monitors.
 package mon
 
 import (
 	"fmt"
 
+	"github.com/rook/rook/pkg/operator/k8sutil"
 	"github.com/rook/rook/pkg/util"
 	"k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/kubernetes/pkg/kubelet/apis"
 )
 
-// detect the nodes that are available for new mons to start.
-func (c *Cluster) getMonNodes() ([]v1.Node, error) {
+// getNodes detect the nodes that are available for new mons to start.
+func (c *Cluster) getNodes() ([]v1.Node, error) {
 	availableNodes, err := c.getAvailableMonNodes()
 	if err != nil {
 		return nil, err
 	}
-	logger.Infof("Found %d running nodes without mons", len(availableNodes))
 
 	if len(availableNodes) == 0 {
 		return nil, fmt.Errorf("no nodes are available for mons")
@@ -48,10 +30,9 @@ func (c *Cluster) getAvailableMonNodes() ([]v1.Node, error) {
 	if err != nil {
 		return nil, err
 	}
-	logger.Debugf("there are %d nodes available for existing %d mons", len(nodes.Items), len(c.clusterInfo.Monitors))
 
 	// get the nodes that have mons assigned
-	nodesInUse, err := c.getNodesWithMons()
+	nodesInUse, err := k8sutil.GetNodesWithApp(c.context.Clientset, c.Namespace, appName)
 	if err != nil {
 		logger.Warningf("could not get nodes with mons. %+v", err)
 		nodesInUse = util.NewSet()
@@ -66,19 +47,4 @@ func (c *Cluster) getAvailableMonNodes() ([]v1.Node, error) {
 	}
 
 	return availableNodes, nil
-}
-
-func (c *Cluster) getNodesWithMons() (*util.Set, error) {
-	options := metav1.ListOptions{LabelSelector: fmt.Sprintf("app=%s", appName)}
-	pods, err := c.context.Clientset.CoreV1().Pods(c.Namespace).List(options)
-	if err != nil {
-		return nil, err
-	}
-	nodes := util.NewSet()
-	for _, pod := range pods.Items {
-		hostname := pod.Spec.NodeSelector[apis.LabelHostname]
-		logger.Debugf("mon pod on node %s", hostname)
-		nodes.Add(hostname)
-	}
-	return nodes, nil
 }
