+++ pkg/operator/k8sutil/nodes.go
@@ -1,0 +1,42 @@
+/*
+Copyright 2018 The Rook Authors. All rights reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+	http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+// Package k8sutil for Kubernetes helpers.
+package k8sutil
+
+import (
+	"fmt"
+
+	"github.com/rook/rook/pkg/util"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/client-go/kubernetes"
+	"k8s.io/kubernetes/pkg/kubelet/apis"
+)
+
+// GetNodesWithApp return nodes with Pods with label `app=$appName` on it
+func GetNodesWithApp(clset kubernetes.Interface, namespace, appName string) (*util.Set, error) {
+	options := metav1.ListOptions{LabelSelector: fmt.Sprintf("app=%s", appName)}
+	pods, err := clset.CoreV1().Pods(namespace).List(options)
+	if err != nil {
+		return nil, err
+	}
+	nodes := util.NewSet()
+	for _, pod := range pods.Items {
+		hostname := pod.Spec.NodeSelector[apis.LabelHostname]
+		nodes.Add(hostname)
+	}
+	return nodes, nil
+}
